requirements.txt
PyYAML>=5.3.1
streamlit==1.7.0
click==7.1.2
https://github.com/dmtlvn/dotdict/archive/refs/heads/main.zip